requirements.txt
pytsk3==20230219
pyewf-python==3.7.2
